# cmd/api/docs/swagger.yaml
definitions:
...
  models.AddAddress:
    properties:
      city:
        type: string
      house_name:
        type: string
      name:
        type: string
      pin:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - city
    - house_name
    - name
    - pin
    - state
    - street
    type: object
...
  models.ChangePassword:
    properties:
      old_password:
        type: string
      password:
        type: string
      re_password:
        type: string
    type: object
...
  models.EditEmail:
    properties:
      email:
        type: string
    type: object
  models.EditName:
    properties:
      name:
        type: string
    type: object
  models.EditPhone:
    properties:
      phone:
        type: string
    type: object
...
  models.UserDetails:
    properties:
      confirmpassword:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /users/cart:
    get:
      description: user can view their cart details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Cart
      tags:
      - User
  /users/cart/remove:
    delete:
      description: user can remove products from their cart
      parameters:
      - description: inventory id
        in: query
        name: inventory
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Remove from Cart
      tags:
      - User
  /users/cart/updateQuantity/minus:
    put:
      consumes:
      - application/json
      description: user can subtract 1 quantity of product from their cart
      parameters:
      - description: inventory id
        in: query
        name: inventory
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Subtract quantity in cart by one
      tags:
      - User
  /users/cart/updateQuantity/plus:
    put:
      consumes:
      - application/json
      description: user can add 1 quantity of product to their cart
      parameters:
      - description: inventory id
        in: query
        name: inventory
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add quantity in cart by one
      tags:
      - User
...
  /users/login:
    post:
      consumes:
      - application/json
      description: user can log in by giving their details
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - User
...
  /users/profile/address:
    get:
      consumes:
      - application/json
      description: user can get all their addresses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Addresses
      tags:
      - User
  /users/profile/address/add:
    post:
      consumes:
      - application/json
      description: user can add their addresses
      parameters:
      - description: address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.AddAddress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Address
      tags:
      - User
  /users/profile/details:
    get:
      consumes:
      - application/json
      description: user can get all their details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get User Details
      tags:
      - User
  /users/profile/edit/email:
    put:
      consumes:
      - application/json
      description: user can change their Email
      parameters:
      - description: edit-email
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.EditEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Edit Email
      tags:
      - User
  /users/profile/edit/name:
    put:
      consumes:
      - application/json
      description: user can change their name
      parameters:
      - description: edit-name
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.EditName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Edit Name
      tags:
      - User
  /users/profile/edit/phone:
    put:
      consumes:
      - application/json
      description: user can change their Phone
      parameters:
      - description: edit-phone
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/models.EditPhone'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Edit Phone
      tags:
      - User
...
  /users/profile/security/change-password:
    put:
      consumes:
      - application/json
      description: user can change their password
      parameters:
      - description: changepassword
        in: body
        name: changepassword
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Change Password
      tags:
      - User
  /users/signup:
    post:
      consumes:
      - application/json
      description: user can signup by giving their details
      parameters:
      - description: signup
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/models.UserDetails'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Signup
      tags:
      - User
...
swagger: "2.0"